# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: ArrControl API
  version: 0.1.0
  license: { name: MIT }
servers: [{ url: /api/v1 }]
security: [{ cookieAuth: [] }]
paths:
...
  /auth/csrf:
    get:
      security: []
      operationId: getAuthCsrf
      description: Issues a CSRF token in both the response and the `__Host-arrcontrol_csrf` Secure, host-only, Path=/, SameSite cookie. Send the value in `X-CSRF-Token` on authentication mutations.
      responses:
        '200':
          description: CSRF token issued; the response must not be cached
          content: { application/json: { schema: { $ref: '#/components/schemas/CsrfToken' } } }
        '500': { $ref: '#/components/responses/Problem' }
...
  /auth/login:
    post:
      security: []
      operationId: loginLocal
      parameters: [{ $ref: '#/components/parameters/CsrfToken' }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/LocalLogin' } } }
      responses:
        '200':
          description: Authenticated; rotating opaque access and refresh cookies were issued
          content: { application/json: { schema: { $ref: '#/components/schemas/AuthSession' } } }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '413': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/Problem' }
  /auth/refresh:
    post:
      security: [{ refreshCookie: [] }]
      operationId: refreshLocalSession
      parameters: [{ $ref: '#/components/parameters/CsrfToken' }]
      responses:
        '200':
          description: Session rotated; replacement access and refresh cookies were issued
          content: { application/json: { schema: { $ref: '#/components/schemas/AuthSession' } } }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/Problem' }
  /auth/logout:
    post:
      security: []
      operationId: logoutLocal
      description: Idempotently revokes the identified session family and clears local authentication cookies.
      parameters: [{ $ref: '#/components/parameters/CsrfToken' }]
      responses:
        '204': { description: Session revoked or already absent }
        '403': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/Problem' }
...
components:
  securitySchemes:
    cookieAuth: { type: apiKey, in: cookie, name: __Host-arrcontrol_session, description: Short-lived opaque access token in a Secure, host-only, Path=/, HttpOnly, SameSite cookie. }
    refreshCookie: { type: apiKey, in: cookie, name: __Host-arrcontrol_refresh, description: Rotating opaque refresh token in a Secure, host-only, Path=/, HttpOnly, SameSite cookie. }
  parameters:
...
    CsrfToken: { name: X-CSRF-Token, in: header, required: true, description: Must exactly match the __Host-arrcontrol_csrf cookie., schema: { type: string, minLength: 43, maxLength: 43 } }
  responses:
    Problem: { description: RFC 9457 problem details, content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } } }
...
  schemas:
...
    CsrfToken: { type: object, additionalProperties: false, required: [token], properties: { token: { type: string, minLength: 43, maxLength: 43 } } }
    LocalLogin: { type: object, additionalProperties: false, required: [email,password], properties: { email: { type: string, format: email, maxLength: 320 }, password: { type: string, format: password, maxLength: 1024, writeOnly: true } } }
    AuthSession: { type: object, additionalProperties: false, required: [userId,email,accessExpiresAt,refreshExpiresAt,csrfToken], properties: { userId: { type: string, format: uuid }, email: { type: string, format: email }, accessExpiresAt: { type: string, format: date-time }, refreshExpiresAt: { type: string, format: date-time }, csrfToken: { type: string, minLength: 43, maxLength: 43 } } }
...
    Problem: { type: object, additionalProperties: false, required: [type,title,status,code,traceId], properties: { type: { type: string, format: uri-reference }, title: { type: string }, status: { type: integer }, detail: { type: string }, traceId: { type: string }, code: { type: string } } }